updated build process to include -ALL option so now can do either single tests like -test sdma OR -test ALL the test -ALL option requires a ALL_test function to be defined, i included this in the imx61.c file, in this file you can add all the tests to be tested sequentially.   Additionally now .conf file needs to be update to include all the driver/test directories.
Also i moved the uart, timer and i2c driver directories into the sdk folder.
not sure if the others in the common are still used for mx53?  so i left them in there but the board_id, fuse, and pmic folders within common i think can be deleted.

updated .conf with the new path for uart and timer folders.

Signed-off-by: Fernando Mireles <f.mireles@freescale.com>